Rename metrics collector to collectors
