+++ nucliadb_core/src/metrics/collectors/noop.rs
@@ -1,4 +1,4 @@
-use crate::metrics::collector::MetricsCollector;
+use crate::metrics::collectors::MetricsCollector;
 use crate::metrics::request_time;
 use crate::NodeResult;
